packages/backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.3
beautifulsoup4==4.12.3
playwright==1.49.1
web3==7.12.0
google-cloud-storage==3.1.0
pyarrow==18.1.0
pytest==8.3.3
